+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Access Alchemy sites' Janeway interface thanks to xterm.js
 * Upgrade alchemy to v1.0.0
 * Add vhost to access log
+* Add http/2 support using the `spdy` module
 
 ## 0.2.0 (2017-09-01)